+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [support-v2]
 
+### Added
+- `VolumeController` to control and manage volume and mute state by multiple `Component`s in a single place
+
 ### Changed
 - Set `UIConfig.playbackSpeedSelectionEnabled` to true per default
 
@@ -13,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Avoid unnecessary animation when `BufferingOverlay` is hidden
 - Avoid unnecessary DOM modification when the text of a `Label` does not change
 - Handling of `null` keys in `SelectBox` (fixes subtitle deselection in IE11)
+- `VolumeToggleButton` interfered player API `setVolume`/`mute`/`unmute` calls
 
 ## [2.19.0] (2018-09-06)